+++ packages/expo-app-metrics/android/src/main/java/expo/modules/appmetrics/logevents/LogEventOptions.kt
@@ -11,6 +11,7 @@ import expo.modules.kotlin.types.OptimizedRecord
  */
 @OptimizedRecord
 data class LogEventOptions(
+  @Field val displayName: String? = null,
   @Field val body: String? = null,
   @Field val attributes: Map<String, Any?>? = null,
   @Field val severity: Severity? = null