+++ packages/expo-app-metrics/ios/EXReactMarker.h
@@ -6,11 +6,7 @@ using namespace facebook::react::ReactMarker;
 NS_SWIFT_NAME(ReactMarker)
 @interface EXAppMetricsReactMarker : NSObject
 
-+ (double)getAppStartupStartTime;
 + (double)getAppStartupEndTime;
 + (double)getRunJSBundleStartTime;
-+ (double)getRunJSBundleEndTime;
-+ (double)getInitReactRuntimeStartTime;
-+ (double)getInitReactRuntimeEndTime;
 
 @end