+++ packages/expo-observe/src/integrations/expo-router/useObserveForRouter.ts
@@ -1,13 +1,13 @@
 import AppMetrics, { type MetricAttributes } from 'expo-app-metrics';
 import { use, useCallback, useEffect, useRef } from 'react';
 
+import { useAssertValueDoesNotChange } from '../../useAssertValueDoesNotChange';
 import { getNavigationMetricParams } from '../navigationConfig';
 import { ObserveRouterIntegrationContext } from './ObserveRouterIntegrationProvider';
 import { emitTTI } from './emitTTI';
 import { getRouterIntegrationConfig, isInitialized } from './init';
 import { buildRoutePattern } from './routeName';
 import { optionalRouter } from './router';
-import { useAssertValueDoesNotChange } from '../../useAssertValueDoesNotChange';
 
 type MarkInteractive = (typeof AppMetrics)['markInteractive'];